moved config back to main package
